--- a/unison-runtime/src/Unison/Runtime/ANF/Serialize.hs
+++ b/unison-runtime/src/Unison/Runtime/ANF/Serialize.hs
@@ -654,6 +654,8 @@ pOpCode op = case op of
   IORI -> 126
   XORI -> 127
   COMI -> 128
+  DRPN -> 129
+  TRNC -> 130
 
 pOpAssoc :: [(POp, Word16)]
 pOpAssoc = map (\op -> (op, pOpCode op)) [minBound .. maxBound]