--- a/unison-runtime/src/Unison/Runtime/MCode.hs
+++ b/unison-runtime/src/Unison/Runtime/MCode.hs
@@ -323,6 +323,7 @@ data UPrim1
   | FLOR -- floor
   | TRNF -- truncate
   | RNDF -- round
+  | TRNC -- truncate
   deriving (Show, Eq, Ord, Enum, Bounded)
 
 data UPrim2
@@ -366,6 +367,7 @@ data UPrim2
   | MAXF -- max
   | MINF -- min
   | CAST -- unboxed runtime type cast (int to nat, etc.)
+  | DRPN -- dropn
   deriving (Show, Eq, Ord, Enum, Bounded)
 
 data BPrim1
@@ -1196,6 +1198,7 @@ emitPOp ANF.ADDI = emitP2 ADDI
 emitPOp ANF.ADDN = emitP2 ADDN
 emitPOp ANF.SUBI = emitP2 SUBI
 emitPOp ANF.SUBN = emitP2 SUBN
+emitPOp ANF.DRPN = emitP2 DRPN
 emitPOp ANF.MULI = emitP2 MULI
 emitPOp ANF.MULN = emitP2 MULN
 emitPOp ANF.DIVI = emitP2 DIVI
@@ -1218,6 +1221,7 @@ emitPOp ANF.INCI = emitP1 INCI
 emitPOp ANF.INCN = emitP1 INCN
 emitPOp ANF.DECI = emitP1 DECI
 emitPOp ANF.DECN = emitP1 DECN
+emitPOp ANF.TRNC = emitP1 TRNC
 emitPOp ANF.TZRO = emitP1 TZRO
 emitPOp ANF.LZRO = emitP1 LZRO
 emitPOp ANF.POPC = emitP1 POPC