--- a/unison-runtime/src/Unison/Runtime/Serialize.hs
+++ b/unison-runtime/src/Unison/Runtime/Serialize.hs
@@ -476,7 +476,10 @@ instance Tag BPrim1 where
   tag2word TLTT = 24
   tag2word DBTX = 25
   tag2word SDBL = 26
-  tag2word RREF = 27
+  tag2word REFN = 27
+  tag2word REFR = 28
+  tag2word RRFC = 29
+  tag2word TIKR = 30
 
   word2tag 0 = pure SIZT
   word2tag 1 = pure USNC
@@ -505,7 +508,10 @@ instance Tag BPrim1 where
   word2tag 24 = pure TLTT
   word2tag 25 = pure DBTX
   word2tag 26 = pure SDBL
-  word2tag 27 = pure RREF
+  word2tag 27 = pure REFN
+  word2tag 28 = pure REFR
+  word2tag 29 = pure RRFC
+  word2tag 30 = pure TIKR
   word2tag n = unknownTag "BPrim1" n
 
 instance Tag BPrim2 where
@@ -535,7 +541,7 @@ instance Tag BPrim2 where
   tag2word IXOT = 23
   tag2word IXOB = 24
   tag2word SDBV = 25
-  tag2word WREF = 26
+  tag2word REFW = 26
 
   word2tag 0 = pure EQLU
   word2tag 1 = pure CMPU
@@ -563,5 +569,5 @@ instance Tag BPrim2 where
   word2tag 23 = pure IXOT
   word2tag 24 = pure IXOB
   word2tag 25 = pure SDBV
-  word2tag 26 = pure WREF
+  word2tag 26 = pure REFW
   word2tag n = unknownTag "BPrim2" n