@ChrisPenner 2024-12-03
@aryairani  Github has been screwing up merge ancestor computation a lot lately, at one point it showed like 40,000 changes and I had to switch the merge base then switch it back to get it to recalculate, so probably at some point in there it listed a change to a ci workflow or something.